split out display and execution state
