+++ .gitignore
@@ -7,3 +7,4 @@ clings
 clings_debug
 .clangd
 compile_commands.json
+.cache
